Some cleanup
This adds [more] to the end of the first page of truncated posts.
This is after feedback from a reader that it wasn't clear that the
front page had a link to more content, and not just a short blurb.
I suspect that will be clearer once there is more content.

It also reorganizes how the date is presented. It struck me as weird
to put the date before the title and then use margins to move the
date after the title. That worked pretty poorly on mobile, so I
took an approach that should work better.